waylandsink: release sink->window refcount in Ready to NULL state instead of finalise()
When using waylandsink as the video sink in wpewebkit the pipeline doesn't
gets destroyed between video playbacks. This means the surface stays obsucring
the browser even once the video playback has ended.

By destorying the surface in READY to NULL state webkit behaves as expected.

Note: somewhere in wpewebkit gstreamer backend might be a better place to
force the pipeline to be destructed and the waylandsink finalise()
function to be called which would achieve the same thing as this patch.

Signed-off-by: Peter Griffin <peter.griffin@linaro.org>